A first pass at using regular expressions to solve the problem that fmatch used to solve.
Passes the first set of tests, and commiting the changes so future work does not accidently cause me to lose this work.

This version handles the wildcard character ? in the haystack string. So "bap" will match with "b?p".

It does this by first converting the needle string "bap" to "[b?][a?][p?]". In regular expressions [] denotes a character set, so each individual character is converted to a 2-character set: its original character plus ?